Fixing bug with wrong type of return value and NaN Has to be double of course if p_x = 0 NaN error

+++ src/main/java/de/uni_passau/fim/se2/litterbox/analytics/metric/Cohesion.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 public class Cohesion <T extends ASTNode> implements MetricExtractor<T>, ScratchVisitor {
     public static final String NAME = "cohesion";
 
-    private int count = 0;
+    private double count = 0;
 
     private double local_cohesion = 0;
